twitter-poller/requirements.txt
twikit>=2.3
requests>=2.31
python-dotenv>=1.0
